backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
google-generativeai==0.3.2
requests==2.31.0
pandas==2.1.4
prophet==1.1.5
python-dotenv==1.0.0

